requirements.txt
clearbit==0.1.7
dj-database-url==0.5.0
Django==2.2.2
django-cors-headers==3.0.2
djangorestframework==3.9.4
djangorestframework-camel-case==1.0.3
djangorestframework-simplejwt==4.3.0
psycopg2==2.8.3
six==1.12.0
PyJWT==1.7.1
requests==2.22.0

